+++ src/api/projects/releases.rs
@@ -8,6 +8,7 @@
 //!
 //! These endpoints are used for querying project releases.
 
+mod create;
 pub mod links;
 mod releases;
 
@@ -15,3 +16,10 @@ pub use self::releases::ProjectReleaseOrderBy;
 pub use self::releases::ProjectReleases;
 pub use self::releases::ProjectReleasesBuilder;
 pub use self::releases::ProjectReleasesBuilderError;
+
+pub use self::create::CreateRelease;
+pub use self::create::CreateReleaseAssetLinks;
+pub use self::create::CreateReleaseAssetLinksBuilder;
+pub use self::create::CreateReleaseAssetLinksBuilderError;
+pub use self::create::CreateReleaseBuilder;
+pub use self::create::CreateReleaseBuilderError;